Move MetadataEditor into a module
