+++ metaspace/webapp/src/modules/MetadataEditor/formStructure.ts
