+++ metaspace/webapp/src/modules/MetadataEditor/index.ts
@@ -1,0 +1,3 @@
+// These are imported directly from router as there is a significant amount of non-common code
+export {default as MetadataEditPage} from './MetadataEditPage.vue';
+export {default as UploadPage} from './UploadPage.vue';
