+++ metaspace/webapp/src/modules/MetadataEditor/inputs/TableInput.vue
