+++ metaspace/webapp/src/modules/MetadataEditor/sections/VisibilityOptionSection.vue
